fix(learner): convert `IsWithinViewport` class into a hook and change initial state from `false` to `true`
